Move Apply button logic to wrapper.
